Update default calibration behaviour
We now always produce a few common zeropoint plots in order to deduce
the zp calibration as the source of any issues.  Also, the applied
calibration is now set by a single config parameter: doApplyUbercal
When True, the meas_mosaic results are applied.  When False, the
FLUXMAG0 header from the processCcd processing is used.