utils: add function to copy a column to a catalog
Not all of the schema entries of the unforced are persisted to the
forced catalogs.  Many of them are useful for the diagnostic plots
which, for the most part, are using the forced catalogs.  This
function is to copy columns from one catalog to another.  Clearly,
the catalogs must be contiguous in memory and have identical entries
in terms of object IDs.