colorAnalysis: add restricted range linear fit plots
The Ivezic "Principal Colors" and their transformations are defined
in terms of SDSS filters and calibrations, hence they are likely to
have offsets and scatter that are not truly representative of an
arbritrary dataset.  We do not currently have a plan for deriving
these transforms for particular datasets, but knowing if the slope,
zeropoint, and scatter vary for linear fits to the regions defined
in Ivezic et al. 2004 (2004AN....325..583I) could still prove a
useful diagnostic, so such fits are included here.