Run comparison: allow independent calibration type
Allow for the visit comparison script to control application of uberCal
results individually for each run being compared.  This could be used,
for example, to compare the same run applying the uberCal to one but
not the other (so the mag diff and astrom plots will show the effective
calibration fields on a per-object basis).