Add classification flag to list of flags to ignore
These plots make use of the star/galaxy classification to label the objects
as one or the other.  If the value of base_ClassificationExtendedness_value
is nan, the source will get labelled a galaxy if not removed.  To eliminate
any confusion, remove any sources for which the classification flag is set.