Add config option whether to write Parquet tables
There may be cases when the user wants plots generated but no tables.
The previous behavior was to write the Parquet tables if the fastparquet
package was available.  This changes allows the user to specify, via the
option -c doWriteParquetTables=False on the command line, that they do
not want tables persisted (regardless of the fastparquet package
availability).