Make columns to copy a config parameter for reuse
This list is used in multiple places, so include it in the config.
The purpose of the list is to copy columns from one source catalog
to another (where both catalogs have identical object lists).  An
example use case is to copy columns from the "meas" catalogs that
into the "forced" catalogs (not all columns and flags from "meas"
get propogated to "forced", and some of these are useful flags for
subselecting qa samples).