Add reference band flags to forced catalogs
It is very useful for qa/debugging to know which band was used as the
reference band for the measurements in forced photometry on the coadds.
These are currently only persisted in the deepCoadd_ref output file.
This commit reads in those catalogs and adds the appropriate columns
to the forced catalogs.  They will therefore be persisted in the
parquet tables for use with the interactive qa analysis tools.