Update flag copying and culling for quiver plot
The quiver plots should be using the unforced catalogs as this is
where the shape measrements are made.  We continue to copy over
the relevant columns to the forced catalogs for the sake of the
interactive plotting analysis (i.e. so that the ellipticities
can be looked at even if only the forced parquet tables have been
loaded).