Ensure dataset loaded is that specified by the input
When looking at the results of a particular rerun, we only want to
consider the datasets in that rerun (i.e. if the dataset does not
exist in the specified rerun, we do not want the butler to follow
the parent chain up to look for another existing one).  The dafPersist
datasetExists function will return True if any dataset is found
along the parent chain, so we replace that here with an explicit
check on the existence of the file in the location specified by
the "input" rerun.