utils.py: add addIntFloatOrStrColumn() function
This function adds a column to the existing schema populated with either
the array of entries provided or assigning each source the scalar provided.
The latter was essentially the functionality provided by the addCcdColumn()
and addPatchColumn() functions, so they have been removed and the calls to
them replace with calls to the new addIntFloatOrStrColumn() funtion.  Note
that the data type for addIntFloatOrStrColumn() has to be one of int, float,
or str.  An exception is thrown otherwise (which also notes that if the user
wishes to add a boolean flag column, they should use the addFlag() function).