Adapt per-field Galacitc Extinction for HSC coeffs
We now have the extiction coefficients for the HSC filters for conversion
from E(B-V) to extinction, A_filter.  Here we update the per-field Galactic
Extinction function to use these values along with the hardwired mean E(B-V)
for the field.  The per-field correction is the fallback if the libraries
for the per-object correction are not available (they are a part of lsst_sims,
so not currently available to the mainline lsst_distrib).