Change names of catalogs for clarity
The catalog that had been called "forced" is actually the catalog
containing the Principal Colors computed in the transformCatalogs
function using the coefficients stored in the ivezicTransformXXX
dicts (where XXX is currently either SDSS or HSC), so renamed
forced to principalColCats and it is now referred to as such within
functions instead of just "catalog".

The catalog forcedCatalogsByFilter, which is a dict containing one
entry per filter (where each "entry" is the forced catalog for that
filter) has been renamed to byFilterForcedCats and is referred to as
byFilterCats within functions instead of just "catalogs".