Add plot of Galactic Extinction applied
Provides a visual of the variation of the Galactic Extinction correction
across the field.  Can be turned off with a config parameter.

This required some updates due to the use of the AllLabeller class
(had not been exercised previously) as there is no benefit in
distinguishing between object types here.

Also updates the filter labels (not relevant for certain axes).