Add wired Principal Color axes to plots
For each principal color, we compute the mean offset and standard deviation
for the wired "canonical" coeffecients to the P2 and P1 axes.  We also fit
for these for the given data.  Both sets of axes are shown on the plot.  A
shift of the P1 line would indicate a fundamental shift in the color
"locus", whereas a shift in the P2 axis indicates a shift in the position
of the highest density of stars which is how the origin of the P1/P2 plane
is selected (i.e. this is likely not a fundamental "offset" in the colors,
but just reflective of the distribution of stars along that particular
pointing.  Such a shift does not affect the mean or standard deviation in
the statistics shown on the histograms).