Update task runners to allow for output redirection
The CompareVisitAnalysisRunner and CompareCoaddAnalysisRunner
TaskRunners allow for a second rerun, the comparison one, to be
specified.  This accommodation broke the ability to specify the
output rerun directory with the --rerun inRerunDir:outRerunDir
format.  This fixes this such that one can now specify all
three rerun directories:
--rerun inRerunDir:outRerunDir --rerun2 in2RerunDir

With the above, the two catalogs in /root/rerun/inRerunDir and
/root/rerun/in2RerunDir will be compared and the plots will be
output in /root/rerun/outRerunDir