Add histograms for full sample (stars + gals)
This was requested a while ago from HSC folk who wanted to get a
visual on the full sample histogram (~luminosity function).

This also includes the "split" objects in the forced vs. unforced
plots (i.e. objects that were classified as one of star/gal in the
unforced, but the other in the forced catalog).

Finally, the list of datasets to plot is sorted by number of objects
such that the smaller samples are plotted on top of the larger ones
for better visibility.