Update application of colorterms
This sets the decision of wether to apply colorterms to the reference
catalog magnitudes as a config parameter (i.e. the user can explicitly
decide not to have them applied, even if they are present in the obs
package).  If the obs package does not supply colorterms, the config
doApplyColorTerms must be set to False (default is True).

It also updates the setting of a "null" colorterm when no colorterm
is to be applied to be able to find the appropriate filter name for
the given reference dataset, noting that a given reference catalog is
not required to follow any standard filter naming convention.  If it
doesn't, a filterMap mapping from the standard to the in-use convention
must be loaded onto the reference object loader.  Therefor, the filter
name for the reference catalog must be obtained from the filter map if
provided, otherwise the generic/canonical name (as definded in the obs
package's mapper) is used.