Force ccd lookup to be based on detector field
In obs_lsst, a given CCD in the detector is uniquely identified by the
dataId's "detector" field.  There are also "raftName" and "detectorName"
field to indicate the raft and position-within-the-raft of a given
detector.  Here, we are only passing in the "detector" field, so we
want the look-up to be based on it only (noting that the dataId
passed into this function is meant to be a sort of generic id for
the whole visit, so the "raftName" and "detectorName" values attached
will not apply to all "detector"s in the visit, so leaving them there
could lead to a failure to located the dataset).