Add axis with units of FWHM for trace plots
Where,

FWHM = 2*sqrt(2*ln(2))*Trace (pixels)

and,

Trace = sqrt(0.5*(Ixx + Iyy)) (pixels)

This is the unit many are most familiar with, so adding an axis with this
unit is helpful for satisfying more user expectations for the values.

It would be nice to have the units in FWHM be arcsec, but this requires
knowing the plate scale for a given instrument/detector.  There is a
plateScale as part of cameraConfig that gets set in the camera.py file
in some obs_* packages, but there does not seem to be consistency in
setting it (e.g. HSC sets plateScale = 1, but the true plate scale is
0.168 "/pix).  If a standardization of the plate scale definition is
established, the FWHM unit can be converted to arcsec.