Add uberCal label to coadd plots and kwargs lists
This is to distinguish between jointcal and meas_mosaic based
uber-calibration of the visit data that was applied to the
single frame exposures before coaddition.

Note that this assumes the presence of the "fcr" flux calibration
datasets from meas_mosaic to identify it as the uber-calibration
method (even though meas_mosaic now also persists a photoCalib
dataset).

To clean things up a bit and ease the addition of new parameters
to all (or many at once) plotting functions, this also introduces
"plotKwargs" dicts (where appropriate) for all in-common plot
function parameters.