Write parquet tables using the butler
This changes the parquet tables to be written using the butler, as
real ParquetTable datasets.  The advantage to this is that the tables
can then be loaded without using pipe_analysis and the Filenamer hack
to get to the filename.  In order for this to work, qa_explorer must be
set up; but if it's not, the parquet writing will just fail with a
warning.