Fix ccd outlining for jointcal output
Jointcal persists photoCalibs for all ccds in a visit, regardless if
it overlaps the tract specified, whereas meas_mosaic only persists
those overlapping the tract.  This led to empty and off axes ccd
outlines being plotted in the comparison plots.

Also added a white dashed line to indicate the tract boundary on
the visit plots (only in the context of having applied an uberCal
which currently operate at the tract level).

Also updated the color scheme for the patches included in the tract
outline panel to use matplotlib's prop_cycle (but swapping the grey
color for gold as grey is the no-data color).