Fix flux columns for Principal Color computations
There were some inconsistencies in doing the transformations to
the Principal Colors for the fit and wired coefficients (we do
this for both PSF mags and CModel mags and they were getting
intermixed).  This is fixed here and we now also use the same
points in all mean and stddev calculations (i.e. those that were
used in the fit and survived the sigma cuts).