Select stats sample based on S/N
Two sets of stats are computed based on "high" and "lower" S/N threshold
levels that are both config parameters.  By default, these are set to
500 and 100, respectively.  These thresholds apply directly for visits,
but for coadds, the threshold gets scaled by square root of the number
of visits.  In either case, if too few objects classified as stars exist
with the configured value, the S/N threshold is decreased by 10 until
a sample with N > config.minHighSampleN (default is 20) is achieved.

The threshold values used are printed to the logs via the Stats object.
An "effective" mag corresponding the S/N cut (computed based as the mean
magnitude of the lower 5% of the S/N > signalToNoiseThreshold subsample)
is also printed on the plots.