Update ref/src angular distance based on uberCals
This updates the "distance" field between the reference catalog and
the source position based on the updated WCS from the uber-calibration
(either jointcal or meas_mosaic).  It also removes the "Match Radius"
label for the reference match plots (as the "matching" there was
done in the stack, we are just reading in those matches in this case).
That is only appropriate when we do a matching ourselves here (e.g. in
the patch overlap and visit/coadd comparison plots).