Add option to print median on plots
This adds the option to print the median in addition to the mean for
the scatter and "sky" plots.  To avoid clutter of all plots where it
is really the mean that is of interest, the median labels are only
added to the "distance" plots which plot the angular distance between
two points of RA and Dec (as this quantity is not symmetric about
zero and is not expected to follow anything approaching a Gaussian
distribution).  Note, however, that the median is at least indicated
on all of the scatter plots as the gray dotted line (as it is the
median that is used to perform the clipping for the statistics).