A few accommodations for DECAM/CFHT data
Namely, that DECam dataIds do not include a "filter" key and the CCD
key is "ccdnum".  Also, CFHT dataIds have a superfluous "extension"
key which can confuse CCD lookup, e.g. when just looping through
a list of CCDs and updating only the CCD key in a dataId.  If CCD
look-up by just the CCD key is desired, make a copy of the dataId
and "pop" the "extention" key (as is done for the LSST raftName and
detectorName keys for the same purpose).