Add first two lsst.verify metrics to colorAnalysis
This adds the framework for adding lsst.verify.metric metrics to
the colorAnalysisTask.  A first two metrics are added and persisted as
a json file.  These metrics, named stellar_locus_width_wPerp and
stellar_locus_width_xPerp, are the width of the "w" and "x" Principal
Color stellar locus (a la Ivezic, Z. et al. 2004, AN 325): the blue
and red sections on the r-i vs g-r plane, respectively.  They are
measured using PSF fluxes.