Add option to add a subdir to the output filepaths
The filepaths are set in the templates to be:
rerunDir/plots/%(filter)s/tract-%(tract)d/[visit-%(visit)d/]
Often, e.g. when testing, one would like to look at a subset of the
ccd/patch unit for visit/coadd level plots, but perhaps do not want
to overwrite the plots from a previous run.  This adds the ability to
add a "subdir" to the filepath such that it will be
added to the end of the canonical filepath.  If only a single unit
(ccd/patch) is requested, the default is to add a subdir of the form
ccd-ccdId/ or patch-N,N/ for visit or coadd level plots respectively.
Otherwise, the subdir can be provided as a command line option:
--subdir.