Add raw and calibrated flux and S/N histograms
This adds plots showing histograms of the fluxes and S/N ratio
distributions for the "raw" intrumental and calibrated fluxes for
all detected sources along with their inversed cumulative fraction.
The was particularly useful when diagnosing the object selection
for PSF modelling (currently based on a flux value that is set as
a config parameter. See DM-17043).  The latter revealed an issue
with the way meas_mosaic was persisting its photoCalib (the errors
were coming out way too high so the S/N ratios were all way smaller
than they should be.  See DM-19828.)