Adapt to astrometry_net import change of DM-21407
As of DM-21407, the required setup of astrometry_net in meas_astrom has
been removed.  It is thus no longer setup here (via the meas_astrom
import).  Since it is only used for compatibility with external catalogs
that have not been ingested in the newer index style (and is currently
unused at all), put the import in a try block and warn if the matched
catalog could not be created due to a failed import.