Add function to read in deepCoadd_obj parquet tables
The postprocess.py script in pipe_tasks outputs a merged multilevel
parquet table with dataset type *_obj (e.g. deepCoadd_obj) consisting
of all relevant dataset tables for the specified input ids (i.e. *_meas,
*_ref, and *_forced_src).  This adds a function to read in these tables
as the input data (rather than the afw SourceCatalogs) to the
coaddAnlysis.py script.

This also updates the pseudo-dynamic creation of the list of columns to
copy from one catalog to another (so they are all available in the plotting
routines) to use schema.extract rather than a search through the list of
all keys if dealing with afw catalogs.