Make use of base_FootprintArea_value in catalogs
Before the existence of this pluggin, a helper function was added
to compute and add the footprint area as a column to the catalogs
called "base_FootprintNPix".  Update to make use of the pluggin-named
column "base_FootprintArea_value" when present (will be with default
configs for all recent, ~post May 2017, catalogs).  Still allow for
computation and addition of this column if it does not exist (i.e.
if reading a very old rerun), but update all usage of "nPix" to use
"Area" instead (to follow convention of pluggin).