Add skyObjects histogram plots
Sky "Objects" corresponding to areas of empty sky (i.e., at positions where
nothing else has been detected) have been added to the multiband coadd
processing (DM-4840) for post-processing assessment of the background
subtraction (i.e. the flux in these empty sky areas should be zero on
average).  Simalarly, sky "Sources" were added to single frame processing
(DM-23078), where we refer to them as sources to distinguish them from the
objects of coadd/multiband processing.  We add here diagnostic plots
showing histograms of various flux measurements (limited to PSF and
circular aperture fluxes as others are explictily ill-defined for blank
sky areas) as well as the sky-distribution of the sky objects/sourcement
placement.

For the coadds, one can also plot the distribution of the sky object
placement on a grayscale image of the coadd with doPlotSkyObjecstSky=True
(the default is False due to the size and somewhat limited utility of the
figure).

This also adds the ability to color-code the camera or tract outlines by
the metric computed per unit (ccd for the former, patch for the latter)
and implements this for the visit/coadd sky source/object plots.