Fix association of flag columns to forced_src catalogs
The coaddAnalysis.py script distinguishes between "unforced" (i.e. the *_meas
dataset type) and "forced" (i.e the *_forced_src dataset type) catalogs.  The
latter catalog does not include some of the flag columns that are useful for
sample sub-selection.  The desired columns are thus copied to the "forced"
catalog within the script.  Most of these columns should be coming from the *_ref
dataset, as it is those that are appropriate for the forced catalogs.  However,
we want the various "calib*" columns to come from the *_meas catalog as these
reflect what actually occurred during SFM per band (which is always what we want
for the calib* columns rather than assigning the specific value associated with
the forced filter to all other filters).  This makes the column copying more
explicit and in accordance with the above desiderata.