Get rid of flagsCat option
In hindsight, this was a flawed way of getting at flags as it conflated
"meas" and "forced_src" values (which can be different for a given
object).  This gets rid of the option to assign a different catalog
from which to select the flags on so that the selection will be made
on the flags based on the catalog being considered itself (having
had the appropriate flags added to it, if necessary, explicitly in
the code).  In the next commit, the logic of how these flags are
associated with a given catalog is fixed.