Limit culling on flags & add flag-set highlighting
When looking at certain sample subsets, in particular a calib_XXX_used subset,
the usual "default QA flag" list should not be selected against since we really
do want to see all of the sources (e.g. that were used in the SFM calibration),
regardless of other flags that may get set (e.g. during coaddition) for a given
source.  However, it would still be good to know if any of these flags were set.
This is done by adding highlights to any such objects (and the highlight indicator
per flag will be added to the legend if any are found).  This makes sure the
subselection based on flags is more explicit for all subsamples and plots and
adds the highlighting of important measurement flags that were not explicitly
culled from the sample.