Add CircAp12 - PSF vs. CircAp12 mag plot
This plot proved very useful for highlighting and debuging issues related
to the aperture correction measurement (see DM-23071).  This also adjusts
the threshold for plotting the running stats as these played a role in
the above diagnosis (the lineweight has been decreased to make them less
imposing when the errorbars are large).