Account for failed star galaxy classification
Our current star/galaxy classification is based on a flux ratio:
Gaussian/PSF for visit-level, CModel/PSF for coadds.  If any of these
flux measurements fails, so will the star/galaxy classification.  The
identification of objects as stars or galaxies in these scripts did not
accommodate nans (these were lumped into the galaxies bin).  This is
fixed here and an "unknown" class is added.

A use case where we have such "unknown" object types is in plotting
the CirpAp - PSF vs. CircAp mags on the coadds.  We do not explicitly
cull on CModel in this case, so some will have failed star/galaxy
classification.