Update to apply distinct external PhotoCalib/SkyWcs
This updates the visitAnalysis and compareVisitAnalysis scripts to be
able to apply distinct external photometric and/or astrometric calibrations
(formerly referred to as "uber calibrations").  Config parameters control
whether and which calibrations to apply so, for example, one could apply
the photometric calibration of fgcmcal but the astrometric calibration from
jointcal.

Various warnings and checks are done to make sure the requested calibrations
are available and self-consistent (e.g. if calibration products were not
found for the same set of CCDs for which src catalogs exist, a warning is
printed).

The plots include a label indicating which calibration is used.  Since this
used up more real estate, I got rid of the "nChild = 0" label as this was
always meant to be temporary (in the transistion period from before the
catalogs were being culled of the "parent" sources that should not be
included in analysis).