Update addition of mag and magErr columns
To accommodate the fake source analysis, columns with magnitudes
and their corresponding errors were added to the catalogs passed
around.  In the original implementation, this also converted the
catalog from an lsst.afw.table.SourceCatalog to a
lsst.afw.table.SimpleCatalog.  This turns out to be a problem as
the latter lacks some of the functionality required for some
applications (e.g. updateCoord on source records).  This changes
the addition of the mag and magErr columns to use the helper
function addIntFloatOrStrColumn() which preserves the catalog
type.