Update catalog updated with external SkyWcs coords
The coordinate update was moved in c5fbce4a but the catalog name
was not updated to be the "calibrated" one.  This update is actually
what revealed the inadvertent change from SourceCatalog to SimpleCatalog
in that there was no updateCoord function on the catalog records. The
latter issue was amended in the previous commit (c3d375d), so all is
now working appropriately.