Adapt all functions to work with DataFrames
We are moving towards reading in the source catalogue data via the
parquet tables produced by the new pipe_tasks postprocessing.py step,
so want all functionality in pipe_analysis to work on pandas DataFrames.
This adaptation is made here.

To allow for processing of repos for which the postprocessing tables
have not been created, allow for (via config doReadParquetTables) the
source catalogs are read in as afw catalogs, but convert them
immediately to pandas DataFrames.

This also adds a readAfwTables function in coaddAnalysis to factor out
this code for reuse by various tasks for(if) reading afwTable catalogs
(i.e. if config.doReadParquetTables = False).  And, in visitAnalysis,
the readCatalogs function is updated for shared use between the two
tasks in that file (VisitAnalysisTask and CompareVisitAnalysisTask).

Note: as of this commit, the reference catalog matching has not been
adapted, and thus will break.  The necessary adaptations will be made
on subsequent commits.