Move metric writing to below plot generation
As of DM-22255, the plot generation is delayed to a final call to
savePlots() in the run method.  As such, the metric measurements do
not get computed and added to the verify job until after this plot
generation function has been run (otherwise we just get a verify
json file with metrics defined, but no measurement values added).