Fix plotting and logic of ccd/patch corner setting
The reworking of the code on DM-22266 resulted in the loss of the
application of the external wcs (i.e. jointcal) calibration (if
performed).  This updates the setting of the ccd corners to use
the correct WCS for the transformation from pixels to sky.  It
also adds the setting of the ccd corners to the match-to-external
reference catalog plots (as they may not be already set if, e.g.
plotMatchesOnly = True).

In the fakes analysis section, the determination of whether a given
source lies within the ccd/patch of interest, the min/max limits were
based on catalog detections, which is not appropriate as these may
not extend to the edges of the ccd/patch.  This is updated here to
use the ccd/patch bounding box (in appropriate RA/Dec coordinates)
as the limits.  Further, in setting the effective area per ccd/patch,
only pixels with tha BAD flag set were being excluded, but an
effective area computation should also be excluding pixels with the
NO_DATA flag set.  This is added here.

It was also discovered here that the setting of the patch corners
in RA/Dec coordinates was flawed (each patch was getting the same
values for the corners).  This is also fixed here.

Finally, some reworking of the returns of readCatalog() are made
to homogenize the structures between the visitAnalysis and
coaddAnalysis versions.