Move getPlotInfo() from fakesAnalysis to plotUtils
Since, as the name describes, this is a function useful for plotting (and is
certainly not specific to fakes analyses), it should live in the plotUtils
module.  This also updates the entries to conform to naming standards in
place throughout this repo (e.g. "camera" refers to the camera objecs, use
cameraName for the name of the camera object).