Remove W504 PEP8 setting from setup.cfg
The project is now recommending we include W503 in the list of exceptions
to PEP8 (https://developer.lsst.io/python/style.html?highlight=w503#exceptions-to-pep-8)
W504 and W503 are incompatible, so the former should have been removed
when the latter was added.  A commit in another ticket (DM-22266) will
go through and clean up all violations of this (and other) style guide
violations.