Fix ability to persist plots to a sub-directory
There is an option in the pipe_analysis scripts for the user to add
a "subdir" to the plot output path.  This is useful in testing when
not wanting to overrite the output files in a given rerun for each
test run (e.g. for comparison purposes).  This functionality was lost
in the DM-22255 refactoring, so it is reinstated (along with template
updates in obs_base and obs_subaru) here.