Persist matched-to-external reference catalogs
The catalogs created by matching the detected sources to a loaded and
denormalized external reference catalog are of use for external QA
analyses and validations.  This adds the persistence of those outputs
as parquet tables in addition to the base catalog-to-parquet persistence
when running with config.doWriteParquetTables=True.

The persisted tables are "denormalized", i.e. contain all fields from
the original source and external catalogs (but with "src_" and "ref_"
prefixes on the column names).