Fix labeling of S/N threshold in colorAnalysis.py
The "high" S/N threshold was being incorrectly printed in the plot
label (although the rough magnitude limit printed was the correct
one).  This labelling is fixed here.  Also, an additional check
on the number of data points going into a polynomial fit is added
in order to provide an informative error message when this case is
encountered.