Add an empty unit string in `addMetricMeasurement`
To create a `Datum` instance of the `lsst_verify` package, the
`quantity` argument should be associated with a unit. Since Rho
Statistics are unitless, we only provide the numerical value and pass an
empty string as its unit. This does not change the default behavior if
`quantity` is an `astropy.units.Quantity` instance, as the `unit`
argument is ignored.