Write functions to load and denormalize matches
This pulls a lot of the functionality from the joinMatchListWithCatalogImpl()
function in meas_algorithms.  Here we add a function to use the information
stored in the persisted "normalized" match catalog in order to load in the
full record for the reference sources within the search radius used in the
SFM calibrations.  The function loadDenormalizeAndUnpackMatches() does a
straight "denormalization" of the persisted normalized match catalog (which
includes only those sources used in the SFM astrometric calibration) and
joins the denormalized (i.e. full object records) of the reference to the
source catalogs and converts and returns it as a pandas DataFrame.

This also adds the function loadReferencesAndMatchToCatalog() to do a
generic match between the loaded reference objects and the source catalog
(regardless of their use or not in any SFM calibration step).  There is
an option to cull the sources based on a list of flags, with the exception
that all sources flagged as having been used in any of the SFM calibration
steps (as indicated by the calib_*_used flag) are retained for further
subselection.