Add functionality to use obj/source parquet tables
The postprocess.py script in pipe_tasks outputs parquet tables that are
based on afw source catalogs.  For coadds, this is a merged multilevel
parquet table with dataset type *_obj (e.g. deepCoadd_obj) consisting
of all relevant dataset tables for the specified input ids (i.e. *_meas,
*_ref, and *_forced_src).  For visits, it is a single level parquet
table with dataset type "source".  This adds a function to read in these
tables as the input data (rather than the afw SourceCatalogs) to the
pipe_analysis scripts.

This also does some clean up and consolidation of the functions that read
in the afw catalogs between the coadd/visitAnalysis scripts to enable reuse
and to reduce code duplication.