Update matching and source (sub)selection plots
This moves from using the loadDenormalizeAndUnpackMatches() function
which simply "denormalizes" the persisted match catalog (which includes
only reference/source pairs that were actually used in the astrometric
calibration during single frame measurement, SFM) to using
loadReferencesAndMatchToCatalog() which loads in the full reference
records for the search radius used in SFM but performs a generic match
to the source catalog, regardless of whether the objects were used in
any calibration step.  Culling of the source catalog based on flags can
be performed prior to matching with the exception that any source used
as a calibarion source will be retained for further (sub)selection and
analysis.