Add Task to compare raw and calibrated WCSs
A Task is added to provide an assessment of the various WCS solutions
for single frame processing.  The goal is to quantify the accuracy of
the initial, raw, WCS with that of the calibrated solutions (whether
from SFP or and external calibration, e.g. jointcal).

A plot showing all WCS solutions as ccd corner outlines in tract pixel
space is also (optionally) created.