Add a check for filter existence in obj parquet tables
Per-patch multilevel parquet tables are created with postprocessing.py for
coadd data where one of the levels is the filter.  As of DM-22266, these are
now the default input catalogs for the pipe_analysis scripts.  In some cases,
e.g. the UDEEP COSMOS field in the HSC RC2 dataset, a given patch may not
contain all filters.  Thus, on trying to read in this patch/filter combo, a
ValueError would be raised as the columns for the non-existant filter do not
exist.  This adds a check on each "obj" table per patch for the existence of
the filter in question.  If it does not exists, the patch is skippped and an
"info" is added to the log.  Note that this modifies the patchDataRefExistsList
in place by removing any dataRefs associated with missing data (i.e. despite
the existence of the obj table for the given patch, there is not data in that
table for the filter under consideration, so it should be removed from the
existence list that was originally based on the on the presence of the obj
parquet tables).