Allow gaia or ps1 ref cat selection for astrometry
We have the ability to use either Gaia or PanSTARRS reference catalogs
for astrometric fitting in jointcal.  As of DM-17597, the default was
set to the Gaia reference catalogs (for astrometric fitting only, the
default for photometry is still PanSTARRS PS1), but obs_subaru was still
overriding to use PS1.  As of DM-25849, this override has been removed so
that HSC processing now uses the jointcal defaults.  The scripts in
pipe_analysis create plots and measure metrics based on a matching of
pipeline sources to reference catalog sources and had PS1 hard-coded as
the reference catalog that gets used in both types of calibration metrics.
This updates the scripts to be able to use either Gaia or PS1 for astrometry
via the config parameter astromRefCat (and defaults to use Gaia).