Add attempt to compute mean epoch for data
Make an attempt to compute the mean epoch for the data under consideration.
For visit-level data, this is simlpy the number stored in the "MDJ" FITS
header.  For coadds, it is the mean of the above for all visits that
contributed to the coadd.  This value is used in the context of correcting
external reference catalogs for proper motion (when available).