Update for treecorr.GGCorrelation API change
The class treecorr.GGCorrelation class no longer has varxi as an attribute,
but now has varxip and varxim, which, according to the documentation at
https://rmjarvis.github.io/TreeCorr/_build/html/index.html represent the
variances of xip (the real part of xi+) and xim (the real part of xi-),
respectively.  It is only xip that is being computed here, so the correct
one to use is varxip.